feat: frontend history store, converter form component
